Feature/update eventspages

+++ MDE.Project.Rosseel_Almanzo/MDE.Project.Rosseel_Almanzo/ViewModels/EventDetailViewModel.cs
@@ -12,6 +12,7 @@ using System.Linq;
 using Newtonsoft.Json.Serialization;
 using MDE.Project.Rosseel_Almanzo.Domain.Services.Interfaces;
 using MDE.Project.Rosseel_Almanzo.Domain.Services;
+using MDE.Project.Rosseel_Almanzo.Infrastructure.Services;
 
 namespace MDE.Project.Rosseel_Almanzo.ViewModels
 {
@@ -29,6 +30,35 @@ namespace MDE.Project.Rosseel_Almanzo.ViewModels
         private ObservableCollection<Domain.Models.Image> images;
         private ObservableCollection<Comment> comments;
         private Comment selectedComment;
+        private bool isAdmin;
+        private string userId;
+        private int attendingUsers;
+
+        public int AttendingUsers
+        {
+            get => attendingUsers;
+            set
+            {
+                attendingUsers = value;
+                RaisePropertyChanged(nameof(AttendingUsers));
+            }
+        }
+
+        public string UserId
+        {
+            get => userId;
+            set => userId = value;
+        }
+
+        public bool IsAdmin
+        {
+            get => isAdmin;
+            set
+            {
+                isAdmin = value;
+                RaisePropertyChanged(nameof(IsAdmin));
+            }
+        }
 
         public Comment SelectedComment
         {
@@ -141,11 +171,15 @@ namespace MDE.Project.Rosseel_Almanzo.ViewModels
             Comments = new ObservableCollection<Comment>();
         }
 
-        public override void Init(object initData)
+        public async override void Init(object initData)
         {
             base.Init(initData);
 
             Id = initData.ToString();
+
+            string admin = await SecureStorage.GetAsync("admin");
+            IsAdmin = bool.Parse(admin);
+            userId = await SecureStorage.GetAsync("token");
 
             GetEventDetails.Execute(null);
         }
@@ -165,6 +199,7 @@ namespace MDE.Project.Rosseel_Almanzo.ViewModels
                     DateEvent = item.DateEvent;
                     Images = item.Images != null ? new ObservableCollection<Domain.Models.Image>(item.Images) : new ObservableCollection<Domain.Models.Image>();
                     Comments = item.Comments != null ? new ObservableCollection<Comment>(item.Comments) : new ObservableCollection<Comment>();
+                    AttendingUsers = item.AttendingUserNames.Count();
                 });
             }
         }
@@ -240,22 +275,60 @@ namespace MDE.Project.Rosseel_Almanzo.ViewModels
             {
                 return new Command(async () =>
                 {
-                    //TODO: Check if comment belongs to logged user!!!!!!!!!!!!!!!!!!!
-
-                    var result = await CoreMethods.DisplayAlert("Delete Comment", "Are u sure u want to delete comment?", "Yes", "Cancel");
-                    if (result)
+                    if (selectedComment.UserId == userId || IsAdmin)
                     {
-                        var deleteResult = await _eventsService.DeleteCommentAsync(Id, selectedComment.Id);
-                        if (deleteResult)
+                        var result = await CoreMethods.DisplayAlert("Delete Comment", "Are u sure u want to delete comment?", "Yes", "Cancel");
+                        if (result)
                         {
-                            Comments.Remove(selectedComment);
-                            await CoreMethods.DisplayAlert("Delete Comment", "Comment succesfull deleted", "Ok");
-                        }
-                        else
-                        {
-                            await CoreMethods.DisplayAlert("Delete Comment", "Delete comment failed!", "Ok");
+                            var deleteResult = await _eventsService.DeleteCommentAsync(Id, selectedComment.Id);
+                            if (deleteResult)
+                            {
+                                Comments.Remove(selectedComment);
+                                await CoreMethods.DisplayAlert("Delete Comment", "Comment succesfull deleted", "Ok");
+                            }
+                            else
+                            {
+                                await CoreMethods.DisplayAlert("Delete Comment", "Delete comment failed!", "Ok");
+                            }
                         }
                     }
+                });
+            }
+        }
+
+        public ICommand DeleteEventCommand
+        {
+            get
+            {
+                return new Command(async () =>
+                {
+                    var result = await _eventsService.DeleteEventAsync(Id);
+                    if (result == "Deleted")
+                    {
+                        await CoreMethods.DisplayAlert("Deleted", "Event succesfull deleted!", "Ok");
+                    }
+                    else
+                    {
+                        await CoreMethods.DisplayAlert("Failed", result, "Ok");
+                    }
+
+                    await CoreMethods.PushPageModel<EventsViewModel>();
+                });
+            }
+        }
+
+        public ICommand SignUpCommand
+        {
+            get
+            {
+                return new Command(async () =>
+                {
+                    var isSigned = await _eventsService.SignUpToEvent(Id, userId);
+                    if (isSigned)
+                    {
+                        AttendingUsers++;
+                        await CoreMethods.DisplayAlert("Succes", "You are succesfully signed up for the event", "Ok");
+                    }                  
                 });
             }
         }
